Slide in on Scroll Functionality
Event listener to window scroll, the function of the event is debounced
so it is not called all the time. On scroll it checks for every slide
photo and if the scroll is past half of the photo, the photo slides in.
If the scroll is past the photo, the photo slides out