Webcam Manipulation Functionality
The functionality starts asking for permision to use the webcam and
seting the video element src to the webcam image value. An event
listener of the 'canplay' event of the video paints the editable webcam
image that paints the webcam image into a canvas and obtains the pixel
values (long array of the image's pixels) and allows its modiication by
iterating each element and altering the pixel values of rgba. The take
photo function plays a snap sound and takes te canvas dataURL and
dowloads it to the user's computer, it also creates an HTML element
showing the img data that was taken