Speech Transcript Functionality
Using the SpeechRecognition JS class the page listens to the microphone
and te 'result' event listener gives what the user is saying. The result
array is processed with maps and converted to a String that is put into
a paragraph. When the user stops speaking, a new paragraph is created.
The 'end' event listener of the recognition class calls the start
function when the user starts speaking again